XML parsing classes refactored.
