+++ Tiralabra_maven/src/main/java/fi/jleh/reittiopas/parser/ReittiopasXMLParser.java
@@ -1,21 +1,20 @@
 package fi.jleh.reittiopas.parser;
 
-import java.io.File;
+import java.io.FileInputStream;
+import java.io.InputStream;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.TreeMap;
 
-import javax.xml.parsers.SAXParser;
-import javax.xml.parsers.SAXParserFactory;
-
-import org.xml.sax.Attributes;
-import org.xml.sax.helpers.DefaultHandler;
+import org.xml.sax.InputSource;
+import org.xml.sax.XMLReader;
+import org.xml.sax.helpers.XMLReaderFactory;
 
 import fi.jleh.reittiopas.exception.XMLParsingException;
 import fi.jleh.reittiopas.model.QuadtreePoint;
 import fi.jleh.reittiopas.model.Service;
 import fi.jleh.reittiopas.model.Station;
-import fi.jleh.reittiopas.model.Stop;
 import fi.jleh.reittiopas.quadtree.BoundingBox;
 import fi.jleh.reittiopas.quadtree.QuadTree;
 import fi.jleh.reittiopas.utils.DataStructuresDto;
@@ -35,87 +34,22 @@ public class ReittiopasXMLParser {
 		long start = System.currentTimeMillis();
 		
 		final List<Station> stationList = new ArrayList<Station>();
-		final TreeMap<Integer, Station> stationMap = new TreeMap<Integer, Station>();
+		final Map<Integer, Station> stationMap = new TreeMap<Integer, Station>();
 		final List<Service> services = new ArrayList<Service>();
 		
 		try {
-			File xmlFile = new java.io.File(pathToFile);
+			InputStream is = new FileInputStream(pathToFile);
+			XMLReader parser = XMLReaderFactory.createXMLReader();
 			
-			SAXParserFactory factory = SAXParserFactory.newInstance();
-			SAXParser parser = factory.newSAXParser();
-			
-			DefaultHandler handler = new DefaultHandler() {
-
-				// Parse values only in one language
-				private boolean synonymOn = false;
-				
-				private Service currentService;
-				
-				public void startElement(String uri, String localName, String qName, Attributes attr) {
-					
-					// Parse stations. Stations under Synonym node are station names in
-					// swedish and they are ignored.
-					if (qName.equalsIgnoreCase("station") && !synonymOn) {
-						int id = Integer.parseInt(attr.getValue("StationId"));
-						
-						Station station = new Station(
-								id,
-								attr.getValue("Name"), 
-								Double.parseDouble(attr.getValue("X")),
-								Double.parseDouble(attr.getValue("Y")), 
-								Integer.parseInt(attr.getValue("city_id")));
-						
-						stationList.add(station);
-						stationMap.put(id, station);
-					}
-					
-					// Start service parsing
-					if (qName.equalsIgnoreCase("service")) {
-						currentService = new Service(Integer.parseInt(attr.getValue("ServiceId")));
-						services.add(currentService);
-					}
-					
-					// Service data is in ServiceNbr node under service node.
-					if (qName.equalsIgnoreCase("ServiceNbr")) {
-						currentService.setCompany(Integer.parseInt(attr.getValue("CompanyId")));
-						currentService.setLineNumber(attr.getValue("Variant"));
-						currentService.setServiceNumber(attr.getValue("ServiceNbr"));
-						currentService.setName(attr.getValue("Name"));
-					}
-					
-					// Parse service stops
-					if (qName.equalsIgnoreCase("stop")) {
-						// Get station for stop
-						Station station = stationMap.get(Integer.parseInt(attr.getValue("StationId")));
-						
-						Stop stop = new Stop(station, attr.getValue("Arrival"), 
-								Integer.parseInt(attr.getValue("Ix")), currentService);
-						
-						currentService.getStops().add(stop);
-						// Add stop to station so we can access lines stopping at station through it
-						station.getStops().add(stop);
-					}
-					
-					if (qName.equalsIgnoreCase("Synonym")) {
-						synonymOn = true;
-					}
-				}
-				
-				public void endElement(String uri, String localName, String qName) {
-					if (qName.equalsIgnoreCase("synonym")) {
-						synonymOn = false;
-					}
-				}
-			};
-			
-			parser.parse(xmlFile, handler);
+			parser.setContentHandler(new ReittiopasXMLHandler(stationList, stationMap, services));
+			parser.parse(new InputSource(is));
 			
 			long end = (System.currentTimeMillis() - start) / 1000;
 			System.out.println("Parsing completed in " + end + " s");
 			System.out.println("Found " + stationList.size() + " stations");
 			System.out.println("Found " + services.size() + " lines");
 		} catch (Exception e) {
-			throw new XMLParsingException("XML parsing failed", e);
+			throw new XMLParsingException("XML parsing failed: " + e, e);
 		}
 		
 		DataStructuresDto dto = new DataStructuresDto();
